Removed Token classes that can never be initialized
They previously existed because Tokenizer.iterchars wanted to return
Token objects, but we can just return the catcode and character. From a
TeX point of view, an Active character is a valid token, but we decided
to encode them as special EscapeSequence's instead, so we can eliminate
the class as well.

The Tokenizer.iterchars function is only used in two places. The first
is the Tokenizer iteration function, which is modified accordingly. It
is also used in TeX.iterchars, which was introduced in cefff9c, but its
use was eliminated in b1d76f9. The function doesn't really make much
sense in the first place and I believe that is why we stopped using it.
I deleted this function completely.