Fix an error resulting from \expandafter being implemented correctly.
Without this PR, you can reproduce the error by replacing nextok with
nexttok in the implementation of \expandafter. The natbib test fails.

The cause was mysterious at first, since nowhere in the source do we use
\expandafter (grepping the whole codebase for expandafter gives only the
`class expandafter(Command):` line).

I figured \expandafter must be invoked in one of the bbl files produced,
which are automatically read by \bibliography. I have not confirmed this
hypothesis, but merely fixed the root cause of the crash.